修改id类型为Long； 增加处理mybatis查询参数工具类ParamMap；

+++ bulk-core/src/test/java/org/bulkframework/core/dao/mybatis/typehandler/FeatureBitTypeHandlerTest.java
@@ -15,7 +15,7 @@
  */
 package org.bulkframework.core.dao.mybatis.typehandler;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertTrue;
 
 import java.util.Calendar;
 
@@ -52,7 +52,7 @@ public class FeatureBitTypeHandlerTest extends AbstractBulkCoreSpringContextTest
 
         user = userDao.insertAndReturnId("typehandler.UserMapper.insertSelective", user);
 
-        Integer id = user.getId();
+        Long id = user.getId();
         assertTrue(id != null && id > 0);
 
         user = userDao.queryOne("typehandler.UserMapper.selectByPrimaryKey", id);
